Rewrite cross-corr
Leave only the introduction to the cross-correlation definition, as that
will be used for optimizations in our implementation.

The FFT implementation description can just redirect to Misko, as we are
basically taking his code anyway and only using it as a baseline.

Remove the attempt at describing the steps from circular
cross-correlation to linear correlation results, just short paragraph
describing it in few words that it is possible, redirecting at misko for
proof. The main part is to describe the algorithm to get the asymptotic
time complexity, as that is important for future analysis why we are
doing what we are doing and how we are doing it.